+++ changelog.md
@@ -21,6 +21,7 @@
 * Set a content-transfer-encoding on multiparts (fixes msglint error)
 * Fix cid generation in MsgHTML (Thanks to @digitalthought)
 * Fix handling of multiple SMTP servers (Thanks to @NanoCaiordo)
+* SMTP->Connect() now supports stream context options (Thanks to @stanislavdavid)
 
 ## Version 5.2.6 (April 11th 2013)
 * Reflect move to PHPMailer GitHub organisation at https://github.com/PHPMailer/PHPMailer